x86_64: refactor thread state creation

+++ oro-arch-x86_64/src/interrupt.rs
@@ -99,7 +99,7 @@ unsafe extern "C" fn isr_sys_timer_rust() -> ! {
 		// If this is `None`, then the kernel is currently running.
 		// Otherwise it's a userspace task that we just jumped from.
 		if let Some(user_task) = scheduler_lock.current_thread().as_ref() {
-			user_task.lock().thread_state_mut().irq_stack_ptr = irq_stack_ptr;
+			user_task.lock().thread_state_mut().irq_stack_ptr = irq_stack_ptr as usize;
 
 			coming_from_user = true;
 		} else {
